Brain_Tumor_Detection/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
tensorflow-cpu==2.15.0
numpy==1.24.3
Pillow==10.1.0
